+++ xmlunit-placeholders/src/main/java/org/xmlunit/placeholder/MatchesRegexPlaceholderHandler.java
@@ -1,3 +1,16 @@
+/*
+  This file is licensed to You under the Apache License, Version 2.0
+  (the "License"); you may not use this file except in compliance with
+  the License.  You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+*/
 package org.xmlunit.placeholder;
 
 import org.xmlunit.XMLUnitException;
@@ -12,6 +25,8 @@ import static org.xmlunit.diff.ComparisonResult.EQUAL;
 
 /**
  * Handler for the {@code matchesRegex()} placeholder keyword.
+ *
+ * @since 2.7.0
  */
 public class MatchesRegexPlaceholderHandler implements PlaceholderHandler {
     private static final String PLACEHOLDER_NAME = "matchesRegex";
@@ -25,7 +40,7 @@ public class MatchesRegexPlaceholderHandler implements PlaceholderHandler {
     public ComparisonResult evaluate(String testText, String... param) {
         if (param.length > 0 && param[0] != null && !param[0].equals("")) {
             try {
-                Pattern pattern = Pattern.compile(param[0].trim());
+                final Pattern pattern = Pattern.compile(param[0].trim());
                 if (testText != null && evaluate(testText.trim(), pattern)) {
                     return EQUAL;
                 }
@@ -36,8 +51,7 @@ public class MatchesRegexPlaceholderHandler implements PlaceholderHandler {
         return DIFFERENT;
     }
 
-    private boolean evaluate(String testText, Pattern pattern) {
-        Matcher matcher = pattern.matcher(testText);
-        return matcher.find();
+    private boolean evaluate(final String testText, final Pattern pattern) {
+        return pattern.matcher(testText).find();
     }
 }