+++ RELEASE_NOTES.md
@@ -1,6 +1,13 @@
 # Release Notes
 
-## XMLUnit for Java 2.6.5 - /not released, yet/
+## XMLUnit for Java 2.7.0 - /not released, yet/
+
+This version contains a backwards incompatible change to the
+`PlaceholderHandler` interface that is part of the experimental
+placeholders module: The `evaluate` method now receives a variable
+number of string arguments in addition to the textual content of the
+element/attribute. This allows placeholders like
+`${xmlunit.matchesRegex("some\s*regex")}`.
 
 * the AssertJ tests now pass on non-English locales as well
   Issue [#180](https://github.com/xmlunit/xmlunit/pull/180)
@@ -8,6 +15,10 @@
 * added a workaround for a binary incompatible change in AssertJ that
   caused xmlunit-assertj to be incompatible with AssertJ 3.15.0
   Issue [#181](https://github.com/xmlunit/xmlunit/issues/181)
+
+* added a `matchesRegex` placeholder
+  PR [#178](https://github.com/xmlunit/xmlunit/issues/178) by
+  [@Jazzyekim](https://github.com/Jazzyekim).
 
 ## XMLUnit for Java 2.6.4 - /Released 2020-03-08/